adding all the latest changes... bad commit

+++ interface/PLTGainCal.h
@@ -15,6 +15,7 @@
 #include "PLTCluster.h"
 #include "PLTPlane.h"
 #include "PLTU.h"
+#include "GetNames.h"
 
 
 class PLTGainCal
@@ -22,18 +23,18 @@ class PLTGainCal
   public:
     PLTGainCal ();
     PLTGainCal (int, bool); // number of ROCs, isExternalFunction
-    PLTGainCal (std::string const, int const NParams = 5);
+    PLTGainCal (std::string const, int const NParams=5);
     ~PLTGainCal ();
 
 
     static int const DEBUGLEVEL = 0;
 
-    void  SetCharge ( PLTHit&, int telescopeID=10);
+    void  SetCharge ( PLTHit&, uint8_t telescopeID=10);
     float GetCharge(int const ch, int const telescopeID, int const roc, int const col, int const row, int adc);
     void  ReadGainCalFile (std::string const GainCalFileName, int roc);
     void  ReadGainCalFile3 (std::string const GainCalFileName);
     void  ReadGainCalFile5 (std::string const GainCalFileName);
-    void  ReadGainCalFileExt (std::string const GainCalFileName, int const roc = 0);
+    void  ReadGainCalFileExt (std::string const GainCalFileName, int const roc=0);
     void  ReadTesterGainCalFile (std::string const GainCalFileName);
 
     void CheckGainCalFile (std::string const GainCalFileName, int const Channel);
